Move serialize deserialize code to type support
