+++ bindings/python/src/subscription/data_reader.rs
@@ -1,8 +1,6 @@
-use dust_dds::serialized_payload::cdr::deserializer::CdrDeserializer;
 use pyo3::{
     exceptions::{PyRuntimeError, PyTypeError},
     prelude::*,
-    types::{PyDict, PyString, PyType},
 };
 
 use crate::{
@@ -18,11 +16,7 @@ use crate::{
         },
         time::Duration,
     },
-    topic_definition::{
-        topic::Topic,
-        type_support::{PythonDdsData, TypeKind},
-    },
-    xtypes::{cdr_deserializer::ClassicCdrDeserializer, endianness},
+    topic_definition::{topic::Topic, type_support::PythonDdsData},
 };
 
 use super::{
@@ -479,74 +473,6 @@ impl DataReader {
     }
 }
 
-fn deserialize_data(
-    py: Python<'_>,
-    type_: Py<PyType>,
-    deserializer: &mut ClassicCdrDeserializer,
-) -> PyResult<Py<PyAny>> {
-    let py_type = type_.bind(py);
-    let object = type_
-        .bind(py)
-        .call_method("__new__", (py_type,), None)?
-        .unbind();
-    let annotations = py_type.getattr("__annotations__")?;
-    let annotation_dict = annotations.downcast::<PyDict>().map_err(PyErr::from)?;
-    for (member_name, member_type) in annotation_dict {
-        let member_name_str = member_name.downcast::<PyString>()?;
-        let member_type_kind = member_type.extract::<TypeKind>()?;
-        match member_type_kind {
-            TypeKind::boolean => {
-                object.setattr(py, member_name_str, deserializer.deserialize_bool()?)?
-            }
-            TypeKind::byte => {
-                object.setattr(py, member_name_str, deserializer.deserialize_u8()?)?
-            }
-            TypeKind::char8 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_char()?)?
-            }
-            TypeKind::char16 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_char()?)?
-            }
-            TypeKind::int8 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_i8()?)?
-            }
-            TypeKind::uint8 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_u8()?)?
-            }
-            TypeKind::int16 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_i16()?)?
-            }
-            TypeKind::uint16 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_u16()?)?
-            }
-            TypeKind::int32 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_i32()?)?
-            }
-            TypeKind::uint32 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_u32()?)?
-            }
-            TypeKind::int64 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_i64()?)?
-            }
-            TypeKind::uint64 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_u64()?)?
-            }
-            TypeKind::float32 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_f32()?)?
-            }
-            TypeKind::float64 => {
-                object.setattr(py, member_name_str, deserializer.deserialize_f64()?)?
-            }
-            TypeKind::float128 => Err(std::io::Error::new(
-                std::io::ErrorKind::Unsupported,
-                "float128 type not yet supported",
-            ))?,
-        };
-    }
-
-    Ok(object)
-}
-
 #[pyclass]
 pub struct Sample {
     sample: dust_dds::subscription::data_reader::Sample<PythonDdsData>,
@@ -557,18 +483,10 @@ pub struct Sample {
 impl Sample {
     #[getter]
     pub fn get_data(&self) -> PyResult<Py<PyAny>> {
-        let python_data = self.sample.data().map_err(into_pyerr)?;
-        let (header, body) = python_data.data.split_at(4);
-        let endianness = match [header[0], header[1]] {
-            endianness::CDR_LE => endianness::CdrEndianness::LittleEndian,
-            endianness::CDR_BE => endianness::CdrEndianness::BigEndian,
-            _ => panic!("Unknown endianness"),
-        };
-        let mut deserializer = ClassicCdrDeserializer::new(body, endianness);
-        Python::with_gil(|py| {
-            let type_ = self.type_.extract(py)?;
-            deserialize_data(py, type_, &mut deserializer)
-        })
+        self.sample
+            .data()
+            .map_err(into_pyerr)?
+            .into_py_object(&self.type_)
     }
 
     #[getter]
