+++ bindings/python/src/xtypes/mod.rs
@@ -1,4 +1,3 @@
 pub mod cdr_deserializer;
 pub mod cdr_serializer;
 pub mod endianness;
-pub mod result;
